python_converter/requirements.txt
fastapi
pandas
chardet
openpyxl
pyarrow
sqlalchemy
pymysql
uvicorn
python-multipart
pydantic
python-dotenv
cryptography
